Refactor PDF gen to its own file
Keeps PDF logic out of main view model for easier rework of report data saving/loading
